extracted scroll to top button
