+++ src/pages/blog/future-of-spam-blocking.js
@@ -8,6 +8,7 @@ import ActionBar from "../../components/ActionBar"
 import TableOfContents from "../../components/TableOfContents"
 import RelatedArticles from "../../components/RelatedArticles"
 import FeaturedImage from "../../components/FeaturedImage"
+import ScrollToTop from "../../components/ScrollToTop"
 
 // ============================================================
 // ARTICLE METADATA - EDIT THIS SECTION FOR EACH NEW ARTICLE
@@ -43,13 +44,8 @@ const articleMetadata = {
 //
 export default function FutureOfSpamBlocking() {
   const [shareCount, setShareCount] = React.useState(0)
-  const [showScrollToTop, setShowScrollToTop] = React.useState(false)
 
-  const scrollToTop = () => {
-    window.scrollTo({ top: 0, behavior: "smooth" })
-  }
-
-  // Share count and scroll handling
+  // Share count handling
   React.useEffect(() => {
     if (typeof window !== "undefined") {
       const currentPath = window.location.pathname
@@ -63,23 +59,6 @@ export default function FutureOfSpamBlocking() {
 
       // Preload share URLs for better performance
       preloadUrls(currentPath)
-
-      // Handle scroll events for scroll-to-top button
-      const timer = setTimeout(() => {
-        const handleScroll = () => {
-          const scrollTop = window.pageYOffset || document.documentElement.scrollTop
-          setShowScrollToTop(scrollTop > 300) // Show button after scrolling 300px
-        }
-
-        window.addEventListener("scroll", handleScroll)
-
-        // Cleanup scroll listener
-        return () => {
-          window.removeEventListener("scroll", handleScroll)
-        }
-      }, 250) // Delay to ensure page is fully loaded
-
-      return () => clearTimeout(timer)
     }
   }, [])
 
@@ -285,32 +264,7 @@ export default function FutureOfSpamBlocking() {
         </div>
 
         {/* Scroll-to-top button */}
-        {showScrollToTop && (
-          <button
-            onClick={scrollToTop}
-            style={{
-              position: "fixed",
-              top: "2rem",
-              left: "calc(50% - 600px - 6rem)",
-              backgroundColor: "transparent",
-              border: "1.5px solid var(--border-color, #eee)",
-              borderRadius: "6px",
-              padding: "6px 8px",
-              display: "flex",
-              alignItems: "center",
-              justifyContent: "center",
-              boxShadow: "0 2px 8px rgba(0, 0, 0, 0.1)",
-              zIndex: 100,
-              cursor: "pointer",
-              transition: "all 0.2s ease",
-              fontSize: "0.9rem",
-              color: "var(--color-text, #333)",
-            }}
-            title="Scroll to top"
-          >
-            ⬆
-          </button>
-        )}
+        <ScrollToTop />
       </div>
     </Wrapper>
   )
