+++ src/pages/blog/template.js
@@ -9,6 +9,7 @@ import ArticleHeader from "../../components/ArticleHeader"
 import ActionBar from "../../components/ActionBar"
 import RelatedArticles from "../../components/RelatedArticles"
 import FeaturedImage from "../../components/FeaturedImage"
+import ScrollToTop from "../../components/ScrollToTop"
 
 // ============================================================
 // ARTICLE METADATA - EDIT THIS SECTION FOR EACH NEW ARTICLE
@@ -31,14 +32,6 @@ const articleMetadata = {
 // ============================================================
 export default function Template() {
   const [shareCount, setShareCount] = React.useState(0)
-  const [showScrollToTop, setShowScrollToTop] = React.useState(false)
-
-  const scrollToTop = () => {
-    window.scrollTo({
-      top: 0,
-      behavior: "smooth",
-    })
-  }
 
   // Preload URLs and load share count when component mounts
   React.useEffect(() => {
@@ -175,19 +168,6 @@ export default function Template() {
         console.log("- testShareCounts() - View all formatting examples")
         console.log("- testUTM() - View captured UTM data")
         console.log("- testShortener() - Test URL shortener functionality")
-
-        // Add scroll listener for scroll-to-top button
-        const handleScroll = () => {
-          const scrollTop = window.pageYOffset || document.documentElement.scrollTop
-          setShowScrollToTop(scrollTop > 300) // Show button after scrolling 300px
-        }
-
-        window.addEventListener("scroll", handleScroll)
-
-        // Cleanup scroll listener
-        return () => {
-          window.removeEventListener("scroll", handleScroll)
-        }
       }, 250) // Delay to ensure page is fully loaded
 
       return () => clearTimeout(timer)
@@ -642,32 +622,7 @@ export default function Template() {
         </div>
 
         {/* Scroll-to-top button */}
-        {showScrollToTop && (
-          <button
-            onClick={scrollToTop}
-            style={{
-              position: "fixed",
-              top: "2rem",
-              left: "calc(50% - 600px - 6rem)",
-              backgroundColor: "transparent",
-              border: "1.5px solid var(--border-color, #eee)",
-              borderRadius: "6px",
-              padding: "6px 8px",
-              display: "flex",
-              alignItems: "center",
-              justifyContent: "center",
-              boxShadow: "0 2px 8px rgba(0, 0, 0, 0.1)",
-              zIndex: 100,
-              cursor: "pointer",
-              transition: "all 0.2s ease",
-              fontSize: "0.9rem",
-              color: "var(--color-text, #333)",
-            }}
-            title="Scroll to top"
-          >
-            ⬆
-          </button>
-        )}
+        <ScrollToTop />
       </div>
     </Wrapper>
   )
